+++ packages/google_maps_flutter/google_maps_flutter/test/fake_google_maps_flutter_platform.dart
@@ -132,8 +132,8 @@ class FakeGoogleMapsFlutterPlatform extends GoogleMapsFlutterPlatform {
     CameraUpdate cameraUpdate, {
     required int mapId,
   }) async {
-    mapInstances[mapId]?.animateCameraConfiguration =
-        CameraUpdateWithConfiguration(
+    mapInstances[mapId]
+        ?.animateCameraConfiguration = CameraUpdateWithConfiguration(
       cameraUpdate: cameraUpdate,
       configuration: null,
     );
@@ -146,8 +146,8 @@ class FakeGoogleMapsFlutterPlatform extends GoogleMapsFlutterPlatform {
     CameraUpdateAnimationConfiguration configuration, {
     required int mapId,
   }) async {
-    mapInstances[mapId]?.animateCameraConfiguration =
-        CameraUpdateWithConfiguration(
+    mapInstances[mapId]
+        ?.animateCameraConfiguration = CameraUpdateWithConfiguration(
       cameraUpdate: cameraUpdate,
       configuration: configuration,
     );
@@ -161,17 +161,14 @@ class FakeGoogleMapsFlutterPlatform extends GoogleMapsFlutterPlatform {
   }) async {}
 
   @override
-  Future<void> setMapStyle(
-    String? mapStyle, {
-    required int mapId,
-  }) async {}
+  Future<void> setMapStyle(String? mapStyle, {required int mapId}) async {}
 
   @override
-  Future<LatLngBounds> getVisibleRegion({
-    required int mapId,
-  }) async {
+  Future<LatLngBounds> getVisibleRegion({required int mapId}) async {
     return LatLngBounds(
-        southwest: const LatLng(0, 0), northeast: const LatLng(0, 0));
+      southwest: const LatLng(0, 0),
+      northeast: const LatLng(0, 0),
+    );
   }
 
   @override
@@ -211,16 +208,12 @@ class FakeGoogleMapsFlutterPlatform extends GoogleMapsFlutterPlatform {
   }
 
   @override
-  Future<double> getZoomLevel({
-    required int mapId,
-  }) async {
+  Future<double> getZoomLevel({required int mapId}) async {
     return 0.0;
   }
 
   @override
-  Future<Uint8List?> takeSnapshot({
-    required int mapId,
-  }) async {
+  Future<Uint8List?> takeSnapshot({required int mapId}) async {
     return null;
   }
 
@@ -316,9 +309,10 @@ class FakeGoogleMapsFlutterPlatform extends GoogleMapsFlutterPlatform {
     if (instance == null) {
       createdIds.add(creationId);
       mapInstances[creationId] = PlatformMapStateRecorder(
-          widgetConfiguration: widgetConfiguration,
-          mapConfiguration: mapConfiguration,
-          mapObjects: mapObjects);
+        widgetConfiguration: widgetConfiguration,
+        mapConfiguration: mapConfiguration,
+        mapObjects: mapObjects,
+      );
       onPlatformViewCreated(creationId);
     }
     return Container();
@@ -340,18 +334,29 @@ class PlatformMapStateRecorder {
     this.mapObjects = const MapObjects(),
     this.mapConfiguration = const MapConfiguration(),
   }) {
-    clusterManagerUpdates.add(ClusterManagerUpdates.from(
-        const <ClusterManager>{}, mapObjects.clusterManagers));
-    groundOverlayUpdates.add(GroundOverlayUpdates.from(
-        const <GroundOverlay>{}, mapObjects.groundOverlays));
+    clusterManagerUpdates.add(
+      ClusterManagerUpdates.from(
+        const <ClusterManager>{},
+        mapObjects.clusterManagers,
+      ),
+    );
+    groundOverlayUpdates.add(
+      GroundOverlayUpdates.from(
+        const <GroundOverlay>{},
+        mapObjects.groundOverlays,
+      ),
+    );
     markerUpdates.add(MarkerUpdates.from(const <Marker>{}, mapObjects.markers));
-    polygonUpdates
-        .add(PolygonUpdates.from(const <Polygon>{}, mapObjects.polygons));
-    polylineUpdates
-        .add(PolylineUpdates.from(const <Polyline>{}, mapObjects.polylines));
+    polygonUpdates.add(
+      PolygonUpdates.from(const <Polygon>{}, mapObjects.polygons),
+    );
+    polylineUpdates.add(
+      PolylineUpdates.from(const <Polyline>{}, mapObjects.polylines),
+    );
     circleUpdates.add(CircleUpdates.from(const <Circle>{}, mapObjects.circles));
-    heatmapUpdates
-        .add(HeatmapUpdates.from(const <Heatmap>{}, mapObjects.heatmaps));
+    heatmapUpdates.add(
+      HeatmapUpdates.from(const <Heatmap>{}, mapObjects.heatmaps),
+    );
     tileOverlaySets.add(mapObjects.tileOverlays);
   }