+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/test/types/map_configuration_test.dart
@@ -15,8 +15,12 @@ void main() {
       webGestureHandling: WebGestureHandling.auto,
       compassEnabled: false,
       mapToolbarEnabled: false,
-      cameraTargetBounds: CameraTargetBounds(LatLngBounds(
-          northeast: const LatLng(30, 20), southwest: const LatLng(10, 40))),
+      cameraTargetBounds: CameraTargetBounds(
+        LatLngBounds(
+          northeast: const LatLng(30, 20),
+          southwest: const LatLng(10, 40),
+        ),
+      ),
       mapType: MapType.normal,
       minMaxZoomPreference: const MinMaxZoomPreference(1.0, 10.0),
       rotateGesturesEnabled: false,
@@ -62,8 +66,9 @@ void main() {
     });
 
     test('handle webGestureHandling', () async {
-      const MapConfiguration diff =
-          MapConfiguration(webGestureHandling: WebGestureHandling.none);
+      const MapConfiguration diff = MapConfiguration(
+        webGestureHandling: WebGestureHandling.none,
+      );
 
       const MapConfiguration empty = MapConfiguration();
       final MapConfiguration updated = diffBase.applyDiff(diff);
@@ -111,10 +116,15 @@ void main() {
     });
 
     test('handle cameraTargetBounds', () async {
-      final CameraTargetBounds newBounds = CameraTargetBounds(LatLngBounds(
-          northeast: const LatLng(55, 15), southwest: const LatLng(5, 15)));
-      final MapConfiguration diff =
-          MapConfiguration(cameraTargetBounds: newBounds);
+      final CameraTargetBounds newBounds = CameraTargetBounds(
+        LatLngBounds(
+          northeast: const LatLng(55, 15),
+          southwest: const LatLng(5, 15),
+        ),
+      );
+      final MapConfiguration diff = MapConfiguration(
+        cameraTargetBounds: newBounds,
+      );
 
       const MapConfiguration empty = MapConfiguration();
       final MapConfiguration updated = diffBase.applyDiff(diff);
@@ -130,8 +140,9 @@ void main() {
     });
 
     test('handle mapType', () async {
-      const MapConfiguration diff =
-          MapConfiguration(mapType: MapType.satellite);
+      const MapConfiguration diff = MapConfiguration(
+        mapType: MapType.satellite,
+      );
 
       const MapConfiguration empty = MapConfiguration();
       final MapConfiguration updated = diffBase.applyDiff(diff);
@@ -148,8 +159,9 @@ void main() {
 
     test('handle minMaxZoomPreference', () async {
       const MinMaxZoomPreference newZoomPref = MinMaxZoomPreference(3.3, 4.5);
-      const MapConfiguration diff =
-          MapConfiguration(minMaxZoomPreference: newZoomPref);
+      const MapConfiguration diff = MapConfiguration(
+        minMaxZoomPreference: newZoomPref,
+      );
 
       const MapConfiguration empty = MapConfiguration();
       final MapConfiguration updated = diffBase.applyDiff(diff);
@@ -165,8 +177,9 @@ void main() {
     });
 
     test('handle rotateGesturesEnabled', () async {
-      const MapConfiguration diff =
-          MapConfiguration(rotateGesturesEnabled: true);
+      const MapConfiguration diff = MapConfiguration(
+        rotateGesturesEnabled: true,
+      );
 
       const MapConfiguration empty = MapConfiguration();
       final MapConfiguration updated = diffBase.applyDiff(diff);
@@ -182,8 +195,9 @@ void main() {
     });
 
     test('handle scrollGesturesEnabled', () async {
-      const MapConfiguration diff =
-          MapConfiguration(scrollGesturesEnabled: true);
+      const MapConfiguration diff = MapConfiguration(
+        scrollGesturesEnabled: true,
+      );
 
       const MapConfiguration empty = MapConfiguration();
       final MapConfiguration updated = diffBase.applyDiff(diff);
@@ -215,8 +229,9 @@ void main() {
     });
 
     test('handle fortyFiveDegreeImageryEnabled', () async {
-      const MapConfiguration diff =
-          MapConfiguration(fortyFiveDegreeImageryEnabled: true);
+      const MapConfiguration diff = MapConfiguration(
+        fortyFiveDegreeImageryEnabled: true,
+      );
 
       const MapConfiguration empty = MapConfiguration();
       final MapConfiguration updated = diffBase.applyDiff(diff);
@@ -312,8 +327,9 @@ void main() {
     });
 
     test('handle myLocationButtonEnabled', () async {
-      const MapConfiguration diff =
-          MapConfiguration(myLocationButtonEnabled: true);
+      const MapConfiguration diff = MapConfiguration(
+        myLocationButtonEnabled: true,
+      );
 
       const MapConfiguration empty = MapConfiguration();
       final MapConfiguration updated = diffBase.applyDiff(diff);
@@ -329,8 +345,10 @@ void main() {
     });
 
     test('handle padding', () async {
-      const EdgeInsets newPadding =
-          EdgeInsets.symmetric(vertical: 1.0, horizontal: 3.0);
+      const EdgeInsets newPadding = EdgeInsets.symmetric(
+        vertical: 1.0,
+        horizontal: 3.0,
+      );
       const MapConfiguration diff = MapConfiguration(padding: newPadding);
 
       const MapConfiguration empty = MapConfiguration();
@@ -465,39 +483,48 @@ void main() {
     });
 
     test('is false with cameraTargetBounds', () async {
-      final CameraTargetBounds newBounds = CameraTargetBounds(LatLngBounds(
-          northeast: const LatLng(55, 15), southwest: const LatLng(5, 15)));
-      final MapConfiguration diff =
-          MapConfiguration(cameraTargetBounds: newBounds);
+      final CameraTargetBounds newBounds = CameraTargetBounds(
+        LatLngBounds(
+          northeast: const LatLng(55, 15),
+          southwest: const LatLng(5, 15),
+        ),
+      );
+      final MapConfiguration diff = MapConfiguration(
+        cameraTargetBounds: newBounds,
+      );
 
       expect(diff.isEmpty, false);
     });
 
     test('is false with mapType', () async {
-      const MapConfiguration diff =
-          MapConfiguration(mapType: MapType.satellite);
+      const MapConfiguration diff = MapConfiguration(
+        mapType: MapType.satellite,
+      );
 
       expect(diff.isEmpty, false);
     });
 
     test('is false with minMaxZoomPreference', () async {
       const MinMaxZoomPreference newZoomPref = MinMaxZoomPreference(3.3, 4.5);
-      const MapConfiguration diff =
-          MapConfiguration(minMaxZoomPreference: newZoomPref);
+      const MapConfiguration diff = MapConfiguration(
+        minMaxZoomPreference: newZoomPref,
+      );
 
       expect(diff.isEmpty, false);
     });
 
     test('is false with rotateGesturesEnabled', () async {
-      const MapConfiguration diff =
-          MapConfiguration(rotateGesturesEnabled: true);
+      const MapConfiguration diff = MapConfiguration(
+        rotateGesturesEnabled: true,
+      );
 
       expect(diff.isEmpty, false);
     });
 
     test('is false with scrollGesturesEnabled', () async {
-      const MapConfiguration diff =
-          MapConfiguration(scrollGesturesEnabled: true);
+      const MapConfiguration diff = MapConfiguration(
+        scrollGesturesEnabled: true,
+      );
 
       expect(diff.isEmpty, false);
     });
@@ -539,15 +566,18 @@ void main() {
     });
 
     test('is false with myLocationButtonEnabled', () async {
-      const MapConfiguration diff =
-          MapConfiguration(myLocationButtonEnabled: true);
+      const MapConfiguration diff = MapConfiguration(
+        myLocationButtonEnabled: true,
+      );
 
       expect(diff.isEmpty, false);
     });
 
     test('is false with padding', () async {
-      const EdgeInsets newPadding =
-          EdgeInsets.symmetric(vertical: 1.0, horizontal: 3.0);
+      const EdgeInsets newPadding = EdgeInsets.symmetric(
+        vertical: 1.0,
+        horizontal: 3.0,
+      );
       const MapConfiguration diff = MapConfiguration(padding: newPadding);
 
       expect(diff.isEmpty, false);