+++ packages/image_picker/image_picker/test/image_picker_test.dart
@@ -483,6 +483,50 @@ void main() {
           expect(response.exception!.message, 'test_error_message');
         });
       });
+
+      group('#pickMultiVideo', () {
+        setUp(() {
+          when(
+            mockPlatform.getMultiVideoWithOptions(options: anyNamed('options')),
+          ).thenAnswer((Invocation _) async => <XFile>[]);
+        });
+
+        test('passes the arguments correctly', () async {
+          final ImagePicker picker = ImagePicker();
+          await picker.pickMultiVideo();
+          await picker.pickMultiVideo(maxDuration: const Duration(seconds: 10));
+          await picker.pickMultiVideo(limit: 5);
+
+          verifyInOrder(<Object>[
+            mockPlatform.getMultiVideoWithOptions(
+              options: argThat(
+                isInstanceOf<MultiVideoPickerOptions>(),
+                named: 'options',
+              ),
+            ),
+            mockPlatform.getMultiVideoWithOptions(
+              options: argThat(
+                isInstanceOf<MultiVideoPickerOptions>().having(
+                  (MultiVideoPickerOptions options) => options.maxDuration,
+                  'maxDuration',
+                  equals(const Duration(seconds: 10)),
+                ),
+                named: 'options',
+              ),
+            ),
+            mockPlatform.getMultiVideoWithOptions(
+              options: argThat(
+                isInstanceOf<MultiVideoPickerOptions>().having(
+                  (MultiVideoPickerOptions options) => options.limit,
+                  'limit',
+                  equals(5),
+                ),
+                named: 'options',
+              ),
+            ),
+          ]);
+        });
+      });
     });
 
     group('#Multi images', () {
@@ -607,9 +651,9 @@ void main() {
                     )
                     .having(
                       (MultiImagePickerOptions options) =>
-                          options.imageOptions.maxHeight,
+                          options.imageOptions.maxWidth,
                       'maxHeight',
-                      equals(20.0),
+                      equals(10.0),
                     )
                     .having(
                       (MultiImagePickerOptions options) =>
@@ -631,9 +675,9 @@ void main() {
                     )
                     .having(
                       (MultiImagePickerOptions options) =>
-                          options.imageOptions.maxHeight,
+                          options.imageOptions.maxWidth,
                       'maxHeight',
-                      equals(20.0),
+                      equals(10.0),
                     )
                     .having(
                       (MultiImagePickerOptions options) =>