+++ packages/image_picker/image_picker_android/example/lib/main.dart
@@ -79,12 +79,11 @@ class _MyHomePageState extends State<MyHomePage> {
   Future<void> _playVideo(XFile? file) async {
     if (file != null && mounted) {
       await _disposeVideoController();
-      late VideoPlayerController controller;
-
-      controller = VideoPlayerController.file(File(file.path));
+      final VideoPlayerController controller = VideoPlayerController.file(
+        File(file.path),
+      );
       _controller = controller;
-      const double volume = 1.0;
-      await controller.setVolume(volume);
+      await controller.setVolume(1.0);
       await controller.initialize();
       await controller.setLooping(true);
       await controller.play();
@@ -95,7 +94,7 @@ class _MyHomePageState extends State<MyHomePage> {
   Future<void> _onImageButtonPressed(
     ImageSource source, {
     required BuildContext context,
-    bool isMultiImage = false,
+    bool allowMultiple = false,
     bool isMedia = false,
   }) async {
     if (_controller != null) {
@@ -103,15 +102,22 @@ class _MyHomePageState extends State<MyHomePage> {
     }
     if (context.mounted) {
       if (_isVideo) {
-        final XFile? file = await _picker.getVideo(
-          source: source,
-          maxDuration: const Duration(seconds: 10),
-        );
-        if (file != null && context.mounted) {
-          _showPickedSnackBar(context, <XFile>[file]);
+        final List<XFile> files;
+        if (allowMultiple) {
+          files = await _picker.getMultiVideoWithOptions();
+        } else {
+          final XFile? file = await _picker.getVideo(
+            source: source,
+            maxDuration: const Duration(seconds: 10),
+          );
+          files = <XFile>[if (file != null) file];
         }
-        await _playVideo(file);
-      } else if (isMultiImage) {
+        if (files.isNotEmpty && context.mounted) {
+          _showPickedSnackBar(context, files);
+          // Just play the first file, to keep the example simple.
+          await _playVideo(files.first);
+        }
+      } else if (allowMultiple) {
         await _displayPickImageDialog(context, true, (
           double? maxWidth,
           double? maxHeight,
@@ -128,7 +134,7 @@ class _MyHomePageState extends State<MyHomePage> {
                 isMedia
                     ? await _picker.getMedia(
                       options: MediaOptions(
-                        allowMultiple: isMultiImage,
+                        allowMultiple: allowMultiple,
                         imageOptions: imageOptions,
                         limit: limit,
                       ),
@@ -163,7 +169,7 @@ class _MyHomePageState extends State<MyHomePage> {
             final XFile? media = _firstOrNull(
               await _picker.getMedia(
                 options: MediaOptions(
-                  allowMultiple: isMultiImage,
+                  allowMultiple: allowMultiple,
                   imageOptions: ImageOptions(
                     maxWidth: maxWidth,
                     maxHeight: maxHeight,
@@ -315,8 +321,7 @@ class _MyHomePageState extends State<MyHomePage> {
     final VideoPlayerController controller = VideoPlayerController.file(
       File(_mediaFileList![index].path),
     );
-    const double volume = 1.0;
-    controller.setVolume(volume);
+    controller.setVolume(1.0);
     controller.initialize();
     controller.setLooping(true);
     controller.play();
@@ -339,7 +344,7 @@ class _MyHomePageState extends State<MyHomePage> {
     if (response.file != null) {
       if (response.type == RetrieveType.video) {
         _isVideo = true;
-        await _playVideo(response.file);
+        await _playVideo(response.files?.firstOrNull);
       } else {
         _isVideo = false;
         setState(() {
@@ -407,8 +412,8 @@ class _MyHomePageState extends State<MyHomePage> {
                 _onImageButtonPressed(ImageSource.gallery, context: context);
               },
               heroTag: 'image0',
-              tooltip: 'Pick Image from gallery',
-              label: const Text('Pick Image from gallery'),
+              tooltip: 'Pick image from gallery',
+              label: const Text('Pick image from gallery'),
               icon: const Icon(Icons.photo),
             ),
           ),
@@ -420,13 +425,12 @@ class _MyHomePageState extends State<MyHomePage> {
                 _onImageButtonPressed(
                   ImageSource.gallery,
                   context: context,
-                  isMultiImage: true,
-                  isMedia: true,
+                  allowMultiple: true,
                 );
               },
-              heroTag: 'multipleMedia',
-              tooltip: 'Pick Multiple Media from gallery',
-              label: const Text('Pick Multiple Media from gallery'),
+              heroTag: 'image1',
+              tooltip: 'Pick multiple images',
+              label: const Text('Pick multiple images'),
               icon: const Icon(Icons.photo_library),
             ),
           ),
@@ -442,9 +446,9 @@ class _MyHomePageState extends State<MyHomePage> {
                 );
               },
               heroTag: 'media',
-              tooltip: 'Pick Single Media from gallery',
-              label: const Text('Pick Single Media from gallery'),
-              icon: const Icon(Icons.photo_library),
+              tooltip: 'Pick item from gallery',
+              label: const Text('Pick item from gallery'),
+              icon: const Icon(Icons.photo_outlined),
             ),
           ),
           Padding(
@@ -455,13 +459,14 @@ class _MyHomePageState extends State<MyHomePage> {
                 _onImageButtonPressed(
                   ImageSource.gallery,
                   context: context,
-                  isMultiImage: true,
+                  allowMultiple: true,
+                  isMedia: true,
                 );
               },
-              heroTag: 'image1',
-              tooltip: 'Pick Multiple Image from gallery',
-              label: const Text('Pick Multiple Image from gallery'),
-              icon: const Icon(Icons.photo_library),
+              heroTag: 'multipleMedia',
+              tooltip: 'Pick multiple items',
+              label: const Text('Pick multiple items'),
+              icon: const Icon(Icons.photo_library_outlined),
             ),
           ),
           Padding(
@@ -472,8 +477,8 @@ class _MyHomePageState extends State<MyHomePage> {
                 _onImageButtonPressed(ImageSource.camera, context: context);
               },
               heroTag: 'image2',
-              tooltip: 'Take a Photo',
-              label: const Text('Take a Photo'),
+              tooltip: 'Take a photo',
+              label: const Text('Take a photo'),
               icon: const Icon(Icons.camera_alt),
             ),
           ),
@@ -485,9 +490,27 @@ class _MyHomePageState extends State<MyHomePage> {
                 _isVideo = true;
                 _onImageButtonPressed(ImageSource.gallery, context: context);
               },
-              heroTag: 'video0',
-              tooltip: 'Pick Video from gallery',
-              label: const Text('Pick Video from gallery'),
+              heroTag: 'video',
+              tooltip: 'Pick video from gallery',
+              label: const Text('Pick video from gallery'),
+              icon: const Icon(Icons.video_file),
+            ),
+          ),
+          Padding(
+            padding: const EdgeInsets.only(top: 16.0),
+            child: FloatingActionButton.extended(
+              backgroundColor: Colors.red,
+              onPressed: () {
+                _isVideo = true;
+                _onImageButtonPressed(
+                  ImageSource.gallery,
+                  context: context,
+                  allowMultiple: true,
+                );
+              },
+              heroTag: 'multiVideo',
+              tooltip: 'Pick multiple videos',
+              label: const Text('Pick multiple videos'),
               icon: const Icon(Icons.video_library),
             ),
           ),
@@ -499,9 +522,9 @@ class _MyHomePageState extends State<MyHomePage> {
                 _isVideo = true;
                 _onImageButtonPressed(ImageSource.camera, context: context);
               },
-              heroTag: 'video1',
-              tooltip: 'Take a Video',
-              label: const Text('Take a Video'),
+              heroTag: 'takeVideo',
+              tooltip: 'Take a video',
+              label: const Text('Take a video'),
               icon: const Icon(Icons.videocam),
             ),
           ),