+++ packages/image_picker/image_picker_linux/example/lib/main.dart
@@ -75,7 +75,7 @@ class _MyHomePageState extends State<MyHomePage> {
   Future<void> _onImageButtonPressed(
     ImageSource source, {
     required BuildContext context,
-    bool isMultiImage = false,
+    bool allowMultiple = false,
     bool isMedia = false,
   }) async {
     if (_controller != null) {
@@ -83,39 +83,49 @@ class _MyHomePageState extends State<MyHomePage> {
     }
     if (context.mounted) {
       if (_isVideo) {
-        final XFile? file = await _picker.getVideo(
-          source: source,
-          maxDuration: const Duration(seconds: 10),
-        );
-        await _playVideo(file);
-      } else if (isMultiImage) {
+        final List<XFile> files;
+        if (allowMultiple) {
+          files = await _picker.getMultiVideoWithOptions();
+        } else {
+          final XFile? file = await _picker.getVideo(
+            source: source,
+            maxDuration: const Duration(seconds: 10),
+          );
+          files = <XFile>[if (file != null) file];
+        }
+        if (files.isNotEmpty && context.mounted) {
+          _showPickedSnackBar(context, files);
+          // Just play the first file, to keep the example simple.
+          await _playVideo(files.first);
+        }
+      } else if (allowMultiple) {
         await _displayPickImageDialog(context, (
           double? maxWidth,
           double? maxHeight,
           int? quality,
         ) async {
           try {
+            final ImageOptions imageOptions = ImageOptions(
+              maxWidth: maxWidth,
+              maxHeight: maxHeight,
+              imageQuality: quality,
+            );
             final List<XFile> pickedFileList =
                 isMedia
                     ? await _picker.getMedia(
                       options: MediaOptions(
-                        allowMultiple: isMultiImage,
-                        imageOptions: ImageOptions(
-                          maxWidth: maxWidth,
-                          maxHeight: maxHeight,
-                          imageQuality: quality,
-                        ),
+                        allowMultiple: allowMultiple,
+                        imageOptions: imageOptions,
                       ),
                     )
                     : await _picker.getMultiImageWithOptions(
                       options: MultiImagePickerOptions(
-                        imageOptions: ImageOptions(
-                          maxWidth: maxWidth,
-                          maxHeight: maxHeight,
-                          imageQuality: quality,
-                        ),
+                        imageOptions: imageOptions,
                       ),
                     );
+            if (pickedFileList.isNotEmpty && context.mounted) {
+              _showPickedSnackBar(context, pickedFileList);
+            }
             setState(() {
               _mediaFileList = pickedFileList;
             });
@@ -136,7 +146,7 @@ class _MyHomePageState extends State<MyHomePage> {
             final XFile? media = _firstOrNull(
               await _picker.getMedia(
                 options: MediaOptions(
-                  allowMultiple: isMultiImage,
+                  allowMultiple: allowMultiple,
                   imageOptions: ImageOptions(
                     maxWidth: maxWidth,
                     maxHeight: maxHeight,
@@ -171,13 +181,12 @@ class _MyHomePageState extends State<MyHomePage> {
                 imageQuality: quality,
               ),
             );
-            setState(() {
-              _setImageFileListFromFile(pickedFile);
-            });
+            if (pickedFile != null && context.mounted) {
+              _showPickedSnackBar(context, <XFile>[pickedFile]);
+            }
+            setState(() => _setImageFileListFromFile(pickedFile));
           } catch (e) {
-            setState(() {
-              _pickImageError = e;
-            });
+            setState(() => _pickImageError = e);
           }
         });
       }
@@ -238,24 +247,34 @@ class _MyHomePageState extends State<MyHomePage> {
         child: ListView.builder(
           key: UniqueKey(),
           itemBuilder: (BuildContext context, int index) {
+            final XFile image = _mediaFileList![index];
             final String? mime = lookupMimeType(_mediaFileList![index].path);
-            return Semantics(
-              label: 'image_picker_example_picked_image',
-              child:
-                  mime == null || mime.startsWith('image/')
-                      ? Image.file(
-                        File(_mediaFileList![index].path),
-                        errorBuilder: (
-                          BuildContext context,
-                          Object error,
-                          StackTrace? stackTrace,
-                        ) {
-                          return const Center(
-                            child: Text('This image type is not supported'),
-                          );
-                        },
-                      )
-                      : _buildInlineVideoPlayer(index),
+            return Column(
+              mainAxisSize: MainAxisSize.min,
+              children: <Widget>[
+                Text(
+                  image.name,
+                  key: const Key('image_picker_example_picked_image_name'),
+                ),
+                Semantics(
+                  label: 'image_picker_example_picked_image',
+                  child:
+                      mime == null || mime.startsWith('image/')
+                          ? Image.file(
+                            File(_mediaFileList![index].path),
+                            errorBuilder: (
+                              BuildContext context,
+                              Object error,
+                              StackTrace? stackTrace,
+                            ) {
+                              return const Center(
+                                child: Text('This image type is not supported'),
+                              );
+                            },
+                          )
+                          : _buildInlineVideoPlayer(index),
+                ),
+              ],
             );
           },
           itemCount: _mediaFileList!.length,
@@ -278,8 +297,7 @@ class _MyHomePageState extends State<MyHomePage> {
     final VideoPlayerController controller = VideoPlayerController.file(
       File(_mediaFileList![index].path),
     );
-    const double volume = 1.0;
-    controller.setVolume(volume);
+    controller.setVolume(1.0);
     controller.initialize();
     controller.setLooping(true);
     controller.play();
@@ -311,8 +329,8 @@ class _MyHomePageState extends State<MyHomePage> {
                 _onImageButtonPressed(ImageSource.gallery, context: context);
               },
               heroTag: 'image0',
-              tooltip: 'Pick Image from gallery',
-              label: const Text('Pick Image from gallery'),
+              tooltip: 'Pick image from gallery',
+              label: const Text('Pick image from gallery'),
               icon: const Icon(Icons.photo),
             ),
           ),
@@ -324,13 +342,12 @@ class _MyHomePageState extends State<MyHomePage> {
                 _onImageButtonPressed(
                   ImageSource.gallery,
                   context: context,
-                  isMultiImage: true,
-                  isMedia: true,
+                  allowMultiple: true,
                 );
               },
-              heroTag: 'multipleMedia',
-              tooltip: 'Pick Multiple Media from gallery',
-              label: const Text('Pick Multiple Media from gallery'),
+              heroTag: 'image1',
+              tooltip: 'Pick multiple images',
+              label: const Text('Pick multiple images'),
               icon: const Icon(Icons.photo_library),
             ),
           ),
@@ -346,9 +363,9 @@ class _MyHomePageState extends State<MyHomePage> {
                 );
               },
               heroTag: 'media',
-              tooltip: 'Pick Single Media from gallery',
-              label: const Text('Pick Single Media from gallery'),
-              icon: const Icon(Icons.photo_library),
+              tooltip: 'Pick item from gallery',
+              label: const Text('Pick item from gallery'),
+              icon: const Icon(Icons.photo_outlined),
             ),
           ),
           Padding(
@@ -359,13 +376,14 @@ class _MyHomePageState extends State<MyHomePage> {
                 _onImageButtonPressed(
                   ImageSource.gallery,
                   context: context,
-                  isMultiImage: true,
+                  allowMultiple: true,
+                  isMedia: true,
                 );
               },
-              heroTag: 'image1',
-              tooltip: 'Pick Multiple Image from gallery',
-              label: const Text('Pick Multiple Image from gallery'),
-              icon: const Icon(Icons.photo_library),
+              heroTag: 'multipleMedia',
+              tooltip: 'Pick multiple items',
+              label: const Text('Pick multiple items'),
+              icon: const Icon(Icons.photo_library_outlined),
             ),
           ),
           if (_picker.supportsImageSource(ImageSource.camera))
@@ -377,8 +395,8 @@ class _MyHomePageState extends State<MyHomePage> {
                   _onImageButtonPressed(ImageSource.camera, context: context);
                 },
                 heroTag: 'image2',
-                tooltip: 'Take a Photo',
-                label: const Text('Take a Photo'),
+                tooltip: 'Take a photo',
+                label: const Text('Take a photo'),
                 icon: const Icon(Icons.camera_alt),
               ),
             ),
@@ -390,9 +408,27 @@ class _MyHomePageState extends State<MyHomePage> {
                 _isVideo = true;
                 _onImageButtonPressed(ImageSource.gallery, context: context);
               },
-              heroTag: 'video0',
-              tooltip: 'Pick Video from gallery',
-              label: const Text('Pick Video from gallery'),
+              heroTag: 'video',
+              tooltip: 'Pick video from gallery',
+              label: const Text('Pick video from gallery'),
+              icon: const Icon(Icons.video_file),
+            ),
+          ),
+          Padding(
+            padding: const EdgeInsets.only(top: 16.0),
+            child: FloatingActionButton.extended(
+              backgroundColor: Colors.red,
+              onPressed: () {
+                _isVideo = true;
+                _onImageButtonPressed(
+                  ImageSource.gallery,
+                  context: context,
+                  allowMultiple: true,
+                );
+              },
+              heroTag: 'multiVideo',
+              tooltip: 'Pick multiple videos',
+              label: const Text('Pick multiple videos'),
               icon: const Icon(Icons.video_library),
             ),
           ),
@@ -405,9 +441,9 @@ class _MyHomePageState extends State<MyHomePage> {
                   _isVideo = true;
                   _onImageButtonPressed(ImageSource.camera, context: context);
                 },
-                heroTag: 'video1',
-                tooltip: 'Take a Video',
-                label: const Text('Take a Video'),
+                heroTag: 'takeVideo',
+                tooltip: 'Take a video',
+                label: const Text('Take a video'),
                 icon: const Icon(Icons.videocam),
               ),
             ),
@@ -492,6 +528,15 @@ class _MyHomePageState extends State<MyHomePage> {
           ],
         );
       },
+    );
+  }
+
+  void _showPickedSnackBar(BuildContext context, List<XFile> files) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text('Picked: ${files.map((XFile it) => it.name).join(',')}'),
+        duration: const Duration(seconds: 2),
+      ),
     );
   }
 }