+++ packages/in_app_purchase/in_app_purchase_storekit/example/integration_test/in_app_purchase_test.dart
@@ -10,8 +10,9 @@ import 'package:integration_test/integration_test.dart';
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  testWidgets('Can create InAppPurchaseStoreKit instance',
-      (WidgetTester tester) async {
+  testWidgets('Can create InAppPurchaseStoreKit instance', (
+    WidgetTester tester,
+  ) async {
     InAppPurchaseStoreKitPlatform.registerPlatform();
     final InAppPurchasePlatform androidPlatform =
         InAppPurchasePlatform.instance;