+++ packages/image_picker/image_picker_windows/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 * Updates minimum supported SDK version to Flutter 3.29/Dart 3.7.
 
+## 0.2.2
+
+* Adds support for `getMultiVideoWithOptions`.
+* Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
+
 ## 0.2.1+1
 
 * Adds pub topics to package metadata.